#6 user, like, cart 컨트롤러 에러 처리

+++ controller/likeController.js
@@ -7,48 +7,77 @@ dotenv.config();
 const addLike = (req, res) => {
     const book_id = req.params.id; // book_id
   
-    let authorization = ensureAuthorization(req);
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES(?, ?)";
-    let values = [authorization.id, book_id];
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else{
+        let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES(?, ?)";
+        let values = [authorization.id, book_id];
 
-    conn.query(sql, values,
-        (err, results ) => {
-            if (err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
-            return res.status(StatusCodes.OK).json(results);
+        conn.query(sql, values,
+            (err, results ) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+                return res.status(StatusCodes.OK).json(results);
         })
+    }
     };
 
 
 const removeLike = (req, res) => {
     const book_id = req.params.id;
 
-   let authorization = ensureAuthorization(req);
-   
-    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
-    let values = [authorization.id, book_id];
+    let authorization = ensureAuthorization(req, res);
 
-    conn.query(sql, values,
-        (err, results ) => {
-            if (err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
-            return res.status(StatusCodes.OK).json(results);
-        })
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else{
+        let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
+        let values = [authorization.id, book_id];
+    
+        conn.query(sql, values,
+            (err, results ) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+                return res.status(StatusCodes.OK).json(results);
+            })
+    }
 }
 
-const ensureAuthorization = (req) => {
-    let receivedJwt = req.headers["authorization"];
-    console.log("received jwt : " ,receivedJwt)
+const ensureAuthorization = (req, res) => {
 
-    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY)
-    console.log(decodedJwt);
+    try{
+        let receivedJwt = req.headers["authorization"];
+        console.log("received jwt : " ,receivedJwt)
 
-    return decodedJwt;
+        let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY)
+        console.log(decodedJwt);
+
+        return decodedJwt;
+    } catch (err) {
+        console.log(err.name);
+        console.log(err.message);
+
+       return err;
+    }
+
 }
 
 module.exports = {
